remove security check.
# Conflicts:
#	Dockerfile

+++ build-and-test.sh
@@ -1,29 +1,33 @@
 #!/bin/bash
 
-set -e
+set -euo pipefail
+
+BUILD="${BUILD:-false}"
 
 echo "Running setup-runtime tests..."
 
 export PIPELINE_FILE="${PIPELINE_FILE:-pipeline.yml}"
 
-docker-compose -p concource-resource -f docker-compose.yml down
-docker-compose -p concource-resource -f docker-compose.yml up -d
-
-docker buildx build \
-  -t localhost:5000/setup-runtime-resource:latest \
-  --progress=plain \
-  --push .
-
-until fly -t test login -c http://localhost:8080 -u test -p test
-do
-  sleep 5
-done
-
 export PIPELINE_NAME="setup-runtime-test"
 
-fly -t test pipelines --json \
-  | jq -r ".[] | select(.name==\"$PIPELINE_NAME\") | .name" \
-  | xargs -n1 -I{} fly -t test destroy-pipeline -p {} -n
+if [ "$BUILD" = "true" ]; then
+  docker-compose -p concource-resource -f docker-compose.yml down
+  docker-compose -p concource-resource -f docker-compose.yml up -d
+
+  docker buildx build \
+    -t localhost:5000/setup-runtime-resource:latest \
+    --progress=plain \
+    --push .
+
+  until fly -t test login -c http://localhost:8080 -u test -p test
+  do
+    sleep 5
+  done
+
+  fly -t test pipelines --json \
+    | jq -r ".[] | select(.name==\"$PIPELINE_NAME\") | .name" \
+    | xargs -n1 -I{} fly -t test destroy-pipeline -p {} -n
+fi
 
 fly -t test set-pipeline -c example/$PIPELINE_FILE -p $PIPELINE_NAME -n\
   --yaml-var "TASK_CONFIG=$(cat example/task.yml)" \
